accounts_microservice/requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
  
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.3
django-cors-headers==3.13.0
djwto==0.0.3
entrypoints==0.4
executing==0.8.3
fastapi==0.79.1
fastjsonschema==2.15.3
flake8==4.0.1
fonttools==4.33.3
h11==0.13.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.2.1
mypy-extensions==0.4.3
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nose==1.3.7
notebook==6.4.11
numpy==1.22.4
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.2
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.0
python-dateutil==2.8.2
pytz==2022.1
boto3==1.24.62
django-storages==1.13.1
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
scipy==1.8.1
selenium==4.1.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.19.1
sympy==1.10.1
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
traitlets==5.2.2.post1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wsproto==1.1.0
# below from service poller
black==22.3.0
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.3
psycopg2==2.9.3
requests==2.27.1
whitenoise==5.3.0
dj-database-url==1.0.0
pika==1.2.0
djangorestframework==3.13.1
